requirements.txt
python-telegram-bot>=20.7
google-generativeai>=0.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=5.0.0
duckduckgo-search>=5.0.0
